refactor: 불필요한 비밀번호 객체 삭제
