+++ backend/text-me/src/main/java/domain/user/entity/User.java
@@ -1,7 +1,7 @@
 package domain.user.entity;
 
 import domain.entity.BaseEntity;
-import domain.security.entity.Password;
+import domain.user.exception.InvalidPasswordException;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
+import java.util.regex.Pattern;
 
 @Entity
 @Getter
@@ -29,13 +30,41 @@ public class User extends BaseEntity {
     private String email;
 
     @Lob
-    @Embedded
     @Column(nullable = false, unique = true)
-    private Password password;
+    private String password;
 
     public User(String name, String email, String password) {
         this.name = name;
         this.email = email;
-        this.password = new Password(password);
+        this.password = password;
+    }
+
+
+    private static void validatePassword(String password) {
+        String REGEX = "^[a-z0-9]{10,20}$";
+        checkSize(password);
+        checkArgument(Pattern.matches(REGEX, password));
+    }
+
+    private static void checkSize(String password) {
+        if (password.isBlank() || password.length() < 8 || password.length() > 12) {
+            throw new IllegalArgumentException("Invalid Password");
+        }
+    }
+
+    private static void checkArgument(boolean matches) {
+        if (!matches) {
+            throw new IllegalArgumentException("Invalid Password");
+        }
+    }
+
+    public void checkSamePassword(String password) {
+        if (!this.password.equals(password)) {
+            throw new InvalidPasswordException("Incorrect Password");
+        }
+    }
+
+    public void setPassword(String encodedPassword) {
+        this.password = encodedPassword;
     }
 }
